+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/upgrade/v5_4_2/PollsPortletIdToDDMPortletIdUpgradeProcess.java
@@ -5,13 +5,12 @@
 
 package com.liferay.dynamic.data.mapping.internal.upgrade.v5_4_2;
 
-import com.liferay.dynamic.data.mapping.internal.upgrade.BasePollsPortletIdUpgradeProcess;
-
 /**
  * @author Carolina Barbosa
  */
 public class PollsPortletIdToDDMPortletIdUpgradeProcess
-	extends BasePollsPortletIdUpgradeProcess {
+	extends com.liferay.dynamic.data.mapping.internal.upgrade.v5_1_4.
+				PollsPortletIdToDDMPortletIdUpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
